[playlists] Move smart playlist validators into a dedicated .cpp
